+++ shared-helpers/__tests__/formKeys.test.ts
@@ -1,5 +1,5 @@
 import { cleanup } from "@testing-library/react"
-import { fieldGroupObjectToArray, prependRoot } from "../src/formKeys"
+import { fieldGroupObjectToArray, prependRoot } from "../src/utilities/formKeys"
 
 afterEach(cleanup)